requirements.txt
click==8.1.8
eval_type_backport==0.2.2
appdirs==1.4.4
multidict==6.4.3
tokenizers==0.21.1
safetensors==0.5.3
aiosignal==1.3.2
charset-normalizer==3.3.2
scipy==1.13.1
PySocks==1.7.1
threadpoolctl==3.5.0
datasets==3.5.0
async-timeout==5.0.1
networkx==3.2.1
GitPython==3.1.44
fsspec==2024.12.0
aiohappyeyeballs==2.6.1
scikit-learn==1.6.1
colorama==0.4.6
MarkupSafe==3.0.2
typing-inspection==0.4.0
filelock==3.17.0
mkl_random==1.2.8
torchaudio==2.3.0
numpy==2.0.1
urllib3==2.3.0
pydantic==2.11.3
tqdm==4.67.1
dill==0.3.8
Jinja2==3.1.6
Brotli==1.0.9
regex==2024.11.6
torch==2.3.0
certifi==2025.1.31
annotated-types==0.7.0
typing_extensions==4.12.2
sentry-sdk==2.26.1
triton==2.3.0
setuptools==75.8.0
transformers==4.51.3
pillow==11.1.0
torchvision==0.18.0
tzdata==2025.2
packaging==24.2
aiohttp==3.11.16
setproctitle==1.3.5
platformdirs==4.3.7
pydantic_core==2.33.1
pandas==2.2.3
docker-pycreds==0.4.0
pyarrow==19.0.0
gitdb==4.0.12
pip==25.0
pytz==2025.2
wandb==0.19.9
six==1.17.0
propcache==0.3.1
sympy==1.13.3
protobuf==5.29.3
xxhash==3.5.0
requests==2.32.3
frozenlist==1.5.0
wheel==0.45.1
python-dateutil==2.9.0.post0
mkl_fft==1.3.11
smmap==5.0.2
idna==3.7
psutil==7.0.0
multiprocess==0.70.16
joblib==1.4.2
gmpy2==2.2.1
PyYAML==6.0.2
mkl-service==2.4.0
yarl==1.20.0
huggingface_hub==0.30.2
mpmath==1.3.0
attrs==25.3.0
